migrated l/rpush, l/rpop, llen with tests
