+++ docs/src/content/docs/commands/LPOP.md
@@ -17,26 +17,32 @@ LPOP key
 | --------- | ------------------------------------------------------------------------------ | ------ | -------- |
 | `key`     | The key of the list from which the first element will be removed and returned. | String | Yes      |
 
+
+
 ## Return Value
 
-| Condition                                      | Return Value                                                                              |
-| ---------------------------------------------- | ----------------------------------------------------------------------------------------- |
-| Command is successful                          | `String` The value of the first element in the list, if the list exists and is not empty. |
-| If the key does not exist or the list is empty | `nil`                                                                                     |
-| Syntax or specified constraints are invalid    | error                                                                                     |
+| Condition                    | Return Value                                         |
+| ---------------------------- | ---------------------------------------------------- |
+| Command is successful        | `String` The value of the first element in the list. |
+| If the key does not exist    | `nil`                                                |
+| The key is of the wrong type | error                                                |
+
+
 
 ## Behavior
 
-- When the `LPOP` command is executed, DiceDB checks if the key exists and is associated with a list. If so, the first element of the list is removed and returned.
-- If the key does not exist or the list is empty, the command returns `nil`.
+- When the `LPOP` command is executed, DiceDB checks if the key exists and is associated with a list. 
+- If the list has elements, the first element is removed and returned.
+- If the key does not exist, the command treats it as an empty list and returns `nil`.
 - If the key exists but is not associated with a list, a `WRONGTYPE` error is returned.
-- If more than one argument is passed, an error is returned.
+- If more than one key is passed, an error is returned.
 
 ## Errors
 
-1. `Wrong type of key`
+1. `Wrong type of value or key`:
 
-When the `LPOP` command is executed for a key that exists but is not associated with a list, an error is returned. This error occurs if the key is associated with a type other than a list, such as a string or set.
+   - Error Message: `(error) WRONGTYPE Operation against a key holding the wrong kind of value`
+   - Occurs if the key exists but is not associated with a list.
 
 ```bash
 127.0.0.1:7379> LPOP mystring
@@ -45,17 +51,8 @@ When the `LPOP` command is executed for a key that exists but is not associated 
 
 2. `Wrong number of arguments`
 
-If the `LPOP` command is executed with more than one key or no key, the following error is returned:
-
-```bash
-127.0.0.1:7379> LPOP
-(error) ERR wrong number of arguments for 'lpop' command
-```
-
-```bash
-127.0.0.1:7379> LPOP mylist secondlist
-(error) ERR wrong number of arguments for 'lpop' command
-```
+   - Error Message: `(error) ERR wrong number of arguments for 'lpop' command`
+   - Occurs if command is executed without any arguments or with 2 or more arguments
 
 ## Example Usage
 
@@ -85,32 +82,27 @@ LPOP mylist
 
 The list `mylist` now contains \["three"\].
 
-### Empty List or Non-Existent Key
+### Non-Existent Key
 
-Returns `(nil)` if the list is empty or the provided key is non-existent
+Returns `(nil)` if the provided key is non-existent
 
 ```bash
 LPOP emptylist
 (nil)
 ```
 
-### Key is Not a List
+### Invalid Usage: Key is Not a List
 
-Setting a key `mystring` with the value `hello`:
+Trying to `LPOP` a key `mystring` which is holding wrong data type `string`.
 
 ```bash
 SET mystring "hello"
 OK
-```
-
-Executing `LPOP` command on any key that is not associated with a List type will result in an error:
-
-```bash
 LPOP mystring
 (error) WRONGTYPE Operation against a key holding the wrong kind of value
 ```
 
-### Wrong Number of Arguments
+### Invalid Usage: Wrong Number of Arguments
 
 Passing more than one key will result in an error:
 
@@ -119,17 +111,10 @@ LPOP mylist secondlist
 (error) ERR wrong number of arguments for 'lpop' command
 ```
 
-### Notes
+## Best Practices
 
-- The key used with the `LPOP` command should be associated with a list to avoid any potential errors.
-- Use the `EXISTS` command to check if a key exists before using `LPOP` to handle cases where the key might not exist.
-- Consider using `LLEN` to check the length of the list if you need to handle empty lists differently.
-
-## Related Commands
-
-- `RPUSH`: Append one or multiple elements to the end of a list.
-- `LPUSH`: Prepend one or multiple elements to the beginning of a list.
-- `RPOP`: Remove and return the last element of a list.
-- `LRANGE`: Get a range of elements from a list.
+- `Check Key Type`: Before using `LPOP`, ensure that the key is associated with a list to avoid errors.
+- `Handle Non-Existent Keys`: Be prepared to handle the case where the key does not exist, as `LPOP` will return `nil` in such scenarios.
+- `Use in Conjunction with Other List Commands`: The `LPOP` command is often used alongside other list commands like `RPUSH`, `LPUSH`, `LLEN`, and `RPOP` to manage and process lists effectively.
 
 By understanding and using the `LPOP` command effectively, you can manage list data structures in DiceDB efficiently, implementing queue-like behaviors and more.
