+++ docs/src/content/docs/commands/RPUSH.md
@@ -13,17 +13,18 @@ RPUSH key value [value ...]
 
 ## Parameters
 
-| Parameter | Description                                                                                                       | Type   | Required |
-| --------- | ----------------------------------------------------------------------------------------------------------------- | ------ | -------- |
-| `key`     | The name of the list where the values will be inserted. If the list does not exist, a new list will be created.   | String | Yes      |
-| `value`   | One or more values to be inserted at the tail of the list. Multiple values can be specified, separated by spaces. | String | Yes      |
+| Parameter          | Description                                                                               | Type   | Required |
+| ------------------ | ----------------------------------------------------------------------------------------- | ------ | -------- |
+| `key`              | The name of the list where values are inserted. If it does not exist, it will be created. | String | Yes      |
+| `value [value...]` | One or more space separated values to be inserted at the tail of the list.                | String | Yes      |
+
 
 ## Return Value
 
-| Condition                                   | Return Value |
-| ------------------------------------------- | ------------ |
-| Command is successful                       | integer      |
-| Syntax or specified constraints are invalid | error        |
+| Condition                                   | Return Value                                   |
+| ------------------------------------------- | ---------------------------------------------- |
+| Command is successful                       | `Integer` - length of the list after execution |
+| Syntax or specified constraints are invalid | error                                          |
 
 ## Behaviour
 
@@ -34,14 +35,15 @@ RPUSH key value [value ...]
 
 ## Errors
 
-1. `Non-List Key`: If the key exists but is not a list, DiceDB returns an error.
+1. `Wrong Number of Arguments`
 
-   - Error Message: `WRONGTYPE Operation against a key holding the wrong kind of value`
-   - Occurs when trying to use the command on a key that does not contain a list.
+    - Error Message: `(error) ERR wrong number of arguments for 'rpush' command`
+    - Occurs if the key parameters is not provided or at least one value is not provided. 
 
-2. `Invalid Syntax`: If the command is not used with the correct syntax, DiceDB returns a syntax error.
-   - Error Message: `ERR wrong number of arguments for 'rpush' command`
-   - Occurs if the command's syntax is incorrect, like wrong number of arguments.
+2. `Wrong Type of Key or Value`: 
+   
+   - Error Message: `(error) WRONGTYPE Operation against a key holding the wrong kind of value`
+   - Occurs if the key exists and is not a list. DiceDB expects the key to either be non-existent or to hold a list data type.
 
 ## Example Usage
 
@@ -60,22 +62,30 @@ Inserting multiple values `world`, `foo`, `bar` into `mylist`
 
 ```bash
 127.0.0.1:7379> RPUSH mylist "world" "foo" "bar"
-(integer) 4
+(integer) 3
 ```
 
-### Invalid Usage
+### Invalid Usage: Key is Not a List
 
-Trying to insert value `val` into a non-list type `mykey`
+Insert the value `value` at the tail of the key `mystring`, which stores a string, not a list.
+
+```shell
+127.0.0.1:7379> SET mystring "not a list"
+OK
+127.0.0.1:7379> RPUSH mystring "value"
+(error) WRONGTYPE Operation against a key holding the wrong kind of value
+```
+
+### Invalid Usage: Wrong Number of Arguments
+
+Calling `RPUSH` without passing any values:
 
 ```bash
-127.0.0.1:7379> SET mykey "notalist"
-127.0.0.1:7379> RPUSH mykey "val"
-(error) ERROR WRONGTYPE Operation against a key holding the wrong kind of value
+RPUSH mylist
+(error) ERR wrong number of arguments for 'rpush' command
 ```
 
-Trying to run `rpush` with only 1 argument
+## Best Practices
 
-```bash
-127.0.0.1:7379> RPUSH mylist
-(error) ERROR wrong number of arguments
-```
+- `Check Key Type`: Before using `RPUSH`, ensure that the key is associated with a list to avoid errors.
+- `Use in Conjunction with Other List Commands`: The `RPUSH` command is often used alongside other list commands like `LLEN`, `LPUSH`, `LPOP`, and `RPOP` to manage and process lists effectively.
