+++ internal/eval/eval.go
@@ -1,4 +1,4 @@
-package eval
+	package eval
 
 import (
 	"bytes"
@@ -1503,195 +1503,6 @@ func evalTOUCH(args []string, store *dstore.Store) []byte {
 	}
 
 	return clientio.Encode(count, false)
-}
-
-func evalLPUSH(args []string, store *dstore.Store) []byte {
-	if len(args) < 2 {
-		return diceerrors.NewErrArity("LPUSH")
-	}
-
-	obj := store.Get(args[0])
-	if obj == nil {
-		obj = store.NewObj(NewDeque(), -1, object.ObjTypeByteList, object.ObjEncodingDeque)
-	}
-
-	// if object is a set type, return error
-	if object.AssertType(obj.TypeEncoding, object.ObjTypeSet) == nil {
-		return diceerrors.NewErrWithFormattedMessage(diceerrors.WrongTypeErr)
-	}
-
-	if err := object.AssertType(obj.TypeEncoding, object.ObjTypeByteList); err != nil {
-		return clientio.Encode(err, false)
-	}
-
-	if err := object.AssertEncoding(obj.TypeEncoding, object.ObjEncodingDeque); err != nil {
-		return clientio.Encode(err, false)
-	}
-
-	store.Put(args[0], obj)
-	for i := 1; i < len(args); i++ {
-		obj.Value.(*Deque).LPush(args[i])
-	}
-
-	deq := obj.Value.(*Deque)
-
-	return clientio.Encode(deq.Length, false)
-}
-
-func evalRPUSH(args []string, store *dstore.Store) []byte {
-	if len(args) < 2 {
-		return diceerrors.NewErrArity("RPUSH")
-	}
-
-	obj := store.Get(args[0])
-	if obj == nil {
-		obj = store.NewObj(NewDeque(), -1, object.ObjTypeByteList, object.ObjEncodingDeque)
-	}
-
-	// if object is a set type, return error
-	if object.AssertType(obj.TypeEncoding, object.ObjTypeSet) == nil {
-		return diceerrors.NewErrWithFormattedMessage(diceerrors.WrongTypeErr)
-	}
-
-	if err := object.AssertType(obj.TypeEncoding, object.ObjTypeByteList); err != nil {
-		return clientio.Encode(err, false)
-	}
-
-	if err := object.AssertEncoding(obj.TypeEncoding, object.ObjEncodingDeque); err != nil {
-		return clientio.Encode(err, false)
-	}
-
-	store.Put(args[0], obj)
-	for i := 1; i < len(args); i++ {
-		obj.Value.(*Deque).RPush(args[i])
-	}
-
-	deq := obj.Value.(*Deque)
-
-	return clientio.Encode(deq.Length, false)
-}
-
-func evalRPOP(args []string, store *dstore.Store) []byte {
-	if len(args) != 1 {
-		return diceerrors.NewErrArity("RPOP")
-	}
-
-	obj := store.Get(args[0])
-	if obj == nil {
-		return clientio.RespNIL
-	}
-
-	// if object is a set type, return error
-	if object.AssertType(obj.TypeEncoding, object.ObjTypeSet) == nil {
-		return diceerrors.NewErrWithFormattedMessage(diceerrors.WrongTypeErr)
-	}
-
-	if err := object.AssertType(obj.TypeEncoding, object.ObjTypeByteList); err != nil {
-		return clientio.Encode(err, false)
-	}
-
-	if err := object.AssertEncoding(obj.TypeEncoding, object.ObjEncodingDeque); err != nil {
-		return clientio.Encode(err, false)
-	}
-
-	deq := obj.Value.(*Deque)
-	x, err := deq.RPop()
-	if err != nil {
-		if errors.Is(err, ErrDequeEmpty) {
-			return clientio.RespNIL
-		}
-		panic(fmt.Sprintf("unknown error: %v", err))
-	}
-
-	return clientio.Encode(x, false)
-}
-
-func evalLPOP(args []string, store *dstore.Store) []byte {
-	// By default we pop only 1
-	popNumber := 1
-
-	// LPOP accepts 1 or 2 arguments only - LPOP key [count]
-	if len(args) < 1 || len(args) > 2 {
-		return diceerrors.NewErrArity("LPOP")
-	}
-
-	// to updated the number of pops
-	if len(args) == 2 {
-		nos, err := strconv.Atoi(args[1])
-		if err != nil {
-			return diceerrors.NewErrWithFormattedMessage(diceerrors.IntOrOutOfRangeErr)
-		}
-		if nos == 0 {
-			// returns empty string if count given is 0
-			return clientio.Encode([]string{}, false)
-		}
-		if nos < 0 {
-			// returns an out of range err if count is negetive
-			return diceerrors.NewErrWithFormattedMessage(diceerrors.ValOutOfRangeErr)
-		}
-		popNumber = nos
-	}
-
-	obj := store.Get(args[0])
-	if obj == nil {
-		return clientio.RespNIL
-	}
-
-	// if object is a set type, return error
-	if object.AssertType(obj.TypeEncoding, object.ObjTypeSet) == nil {
-		return diceerrors.NewErrWithFormattedMessage(diceerrors.WrongTypeErr)
-	}
-
-	if err := object.AssertType(obj.TypeEncoding, object.ObjTypeByteList); err != nil {
-		return clientio.Encode(err, false)
-	}
-
-	if err := object.AssertEncoding(obj.TypeEncoding, object.ObjEncodingDeque); err != nil {
-		return clientio.Encode(err, false)
-	}
-
-	deq := obj.Value.(*Deque)
-
-	// holds the elements popped
-	var elements []string
-	for iter := 0; iter < popNumber; iter++ {
-		x, err := deq.LPop()
-		if err != nil {
-			if errors.Is(err, ErrDequeEmpty) {
-				break
-			}
-			panic(fmt.Sprintf("unknown error: %v", err))
-		}
-		elements = append(elements, x)
-	}
-
-	if len(elements) == 0 {
-		return clientio.RespNIL
-	}
-
-	if len(elements) == 1 {
-		return clientio.Encode(elements[0], false)
-	}
-
-	return clientio.Encode(elements, false)
-}
-
-func evalLLEN(args []string, store *dstore.Store) []byte {
-	if len(args) != 1 {
-		return diceerrors.NewErrArity("LLEN")
-	}
-
-	obj := store.Get(args[0])
-	if obj == nil {
-		return clientio.Encode(0, false)
-	}
-
-	if err := object.AssertTypeAndEncoding(obj.TypeEncoding, object.ObjTypeByteList, object.ObjEncodingDeque); err != nil {
-		return err
-	}
-
-	deq := obj.Value.(*Deque)
-	return clientio.Encode(deq.Length, false)
 }
 
 func evalFLUSHDB(args []string, store *dstore.Store) []byte {
