+++ internal/worker/cmd_meta.go
@@ -63,6 +63,11 @@ const (
 	CmdJSONToggle    = "JSON.TOGGLE"
 	CmdJSONNumIncrBY = "JSON.NUMINCRBY"
 	CmdJSONNumMultBY = "JSON.NUMMULTBY"
+	CmdLPush         = "LPUSH"
+	CmdRPush         = "RPUSH"
+	CmdLPop          = "LPOP"
+	CmdRPop          = "RPOP"
+	CmdLLEN          = "LLEN"
 )
 
 // Multi-shard commands.
@@ -310,6 +315,21 @@ var CommandsMeta = map[string]CmdMeta{
 		CmdType: SingleShard,
 	},
 	CmdJSONNumMultBY: {
+		CmdType: SingleShard,
+	},
+	CmdLPush: {
+		CmdType: SingleShard,
+	},
+	CmdRPush: {
+		CmdType: SingleShard,
+	},
+	CmdLPop: {
+		CmdType: SingleShard,
+	},
+	CmdRPop: {
+		CmdType: SingleShard,
+	},
+	CmdLLEN: {
 		CmdType: SingleShard,
 	},
